Creatinine Normalisation

Creatinine Normalisation (CN) is a useful method much like region of interest normalisation that can normalise spectra based on the total area of the creatinine signal at the chemical shift 3.05ppm.

creNorm(X, ppm = NULL, cre3 = c(3, 3.1), cre4 = c(4, 4.1), err = 5)

Arguments

X
The spectra intended to be normalised. Can either be a single spectrum in the form of a numerical array or multiple spectra in a numerical matrix with the rows being the spectra/samples and the columns being the ppm variables

ppm
A numerical array holding the chemical shift values of the X matrix. Only necessary when X is an array, not when X is a matrix

cre3
A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 3.05 starts and ends.

cre4
A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 4.05 starts and ends.

err
The level of error given when calculating the creatinine peak ratios. interperted as a percentage (i.e., 5 = 5%)

Value

This function assigns the normalised X argument (as X_cre) and the calculated dilution factors (as dilf_cre) to the global environment.

Details

creNorm() works by dividing each element in a row with the sum of the values from its Creatinine signal.

See also

Author

kylebario1@gmail.com

Examples

# When X contains multiple spectra, ppm is not required
data(X, ppm)
creNorm(X)

#> Calculating Dilfs... Done.
#> Checking creatinine peak ratios... spec 1 2 are outside error limits.
#> Refer to Df_cre for more information.
#> Normalising X... Done.
#> 

cat(dilf_cre)

#> 28067256 207466831

# When X has only one spectrum, ppm is required
data(X, ppm)
creNorm(X[1,], ppm)

#> Calculating Dilfs... Done.
#> Checking creatinine peak ratio... The provided spectra is outside of the error limit.
#> Refer to Df_cre for more information.
#> Normalising X... Done.
#> 

cat(dilf_cre)

#> 28067256