Quantile/PQN Normalisation

A method of NMR spectral normalisation where the maximum intensities

q2Norm(X, uv_used = "mode")

Arguments

X
A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables. Cannot normalise only as single spectrum, X must be a matrix.

uv_used
The univariate measure used to calculate the dilution coefficient. mode or median are accepted

Value

This function assigns the normalised X argument (as X_q2) and the calculated dilution factors (as dilf_q2) to the global environment.

Details

The method orders spectral intensities smallest to largest and finds the mean of each 'quantile' (i.e., the mean of the maximum intensities, the mean of the 2nd maximum intensities). These means are then reassigned back to the ppm that first housed the maximum, 2nd maximum, etc., and PQN is performed on each spectrum using its quantile normalised self. The calculated dilution factor is then used to normalise the original spectrum to reduce manipulation of peak shapes.

See also

Author

kylebario1@gmail.com

Examples

data(X)
q2Norm(X, uv_used = 'mode')

#> Calculating Quantile Means... Done.
#> Reassigning Intensity Values... Done.
#> Calculating Quotients... Using the Mode... Done.
#> Normalising X... Done.
#> 

cat(dilf_q2)

#> 0.4546274 1.721394